Discover Weekly Scraper
This scrapes your Discover Weekly after you've listened to it and liked songs. The csv file generated will display song metrics and whether you liked the song or not. It looks at whether the song is in your Liked Songs or not to determine this. This is generally successful, as Discover Weekly typically shows you songs you've never listened to. I've come to learn this is not always the case in about 3 instances over 50 weeks.